Improve formspec escaping
Rather than escaping fields when they are saved,
fields are only escaped when they are used.
While the former method might seem to be more
efficient, it does make the plain text save
file much more messy, and often means that fields
are escaped when they do not need to be. Not only
that, but it does not account for the possibility
of somebody modifying the save file manually,
which while not recommended, is far from impossible.