Allow disabling formspec background
* Formspec background color and image can be
disabled with .pos background false. This is
useful for people who might use the mod with
a server or game lacking Minetest Game's
default background image that the mod utilizes
by default.
* Added .pos help
* Refactored handle_field logic to work with
checks for lists beginning with __
* Lists cannot begin with "__" as it is used
by the mod to store arbitrary data (such as
whether or not the background is enabled)