Backwards compatibility with old versions
Add __storage_version to the modstorage
representing the version of the storage
mechanism used to save the waypoint data.
This means that old installations where
waypoints did not store an HUD toggle or
the added or modified keys will still
function normally

Also, make get_listnames exclude keys that
begin with "__", a forgotten piece of a
previous commit allowing the formspec
background to be toggled on and off.